// backend/src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  ADMIN: 'ADMIN',
  STUDENT: 'STUDENT',
  TEACHER: 'TEACHER',
  REGISTRAR: 'REGISTRAR',
  JANITOR: 'JANITOR',
  AUDITOR: 'AUDITOR',
  GUARD: 'GUARD',
  DEAN: 'DEAN',
  PRINCIPAL: 'PRINCIPAL'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const TeacherDetailsType = {
  MASTERS: 'MASTERS',
  PHD: 'PHD',
  DOCTORATE: 'DOCTORATE'
} as const

export type TeacherDetailsType = (typeof TeacherDetailsType)[keyof typeof TeacherDetailsType]


export const SchoolYearStatus = {
  UPCOMING: 'UPCOMING',
  ONGOING: 'ONGOING',
  COMPLETED: 'COMPLETED'
} as const

export type SchoolYearStatus = (typeof SchoolYearStatus)[keyof typeof SchoolYearStatus]


export const RoomType = {
  CLASSROOM: 'CLASSROOM',
  FACULTY_ROOM: 'FACULTY_ROOM',
  GYMNASIUM: 'GYMNASIUM',
  COMPUTER_LABORATORY: 'COMPUTER_LABORATORY',
  SCIENCE_LABORATORY: 'SCIENCE_LABORATORY'
} as const

export type RoomType = (typeof RoomType)[keyof typeof RoomType]


export const RoomStatus = {
  OPEN: 'OPEN',
  CLOSED: 'CLOSED',
  MAINTENANCE: 'MAINTENANCE'
} as const

export type RoomStatus = (typeof RoomStatus)[keyof typeof RoomStatus]


export const AttendanceStatus = {
  PRESENT: 'PRESENT',
  ABSENT: 'ABSENT',
  LATE: 'LATE',
  EXCUSED: 'EXCUSED'
} as const

export type AttendanceStatus = (typeof AttendanceStatus)[keyof typeof AttendanceStatus]